refactor/merge couple more files in api
